refactor: Rename min_partition_after to min_partition and min_partition_before to min_partition_huristic
